<feat> : 회원탈퇴 기능 추가
<feat> : 회원탈퇴 기능 추가
- 마이페이지 > 회원탈퇴 탭
- 1차조건 유의사항 체크, 2차조건 현재 비밀번호 입력 후 탈퇴 진행

+++ src/main/java/com/project/compagnoserver/service/UserService.java
@@ -53,8 +53,15 @@ public class UserService {
   }
 
   // 회원 탈퇴
-    public void deleteUser(String id) {
-        userDao.deleteUserInfo(id);
+    public int deleteUser(String id, String password, PasswordEncoder encoder) {
+      User user = userDao.findById(id).orElse(null);
+      if(user!= null && encoder.matches(password, user.getPassword())) {
+          userDao.deleteUserInfo(id);
+          // 1 리턴하면 회원탈퇴 성공
+          return 1;
+      }
+        // 0 리턴하면 회원정보 틀려서 탈퇴 실패
+        return 0;
     }
 
 }
